moved merge_cdb to utils and created test_cdb_utils
